libs/neuronet: Implemented a special packet header system.
This special packet header system sends the size of
the packet right before the packet itself is sent, meant
to ease defragmenting packets with protocols that don't
support it. This system should only be used when the
client and server are both set to work with it. This
feature is disabled by default.