libs/neuronet: Rewrote the whole library almost from scratch.
Total redesign of the module, the functions were reunited in
more proper source files.
Dropped the use of callbacks altogether. Now NNet_Poll returns a state
which can be used to know what happens. Crude version that wasn't tested
much and cross porting is coded but not activated, this will be the next
commit.