libs/neuronet: Added flag to avoid the port to be locked.
Under normal circumstances, when a server secures a port for
data transmissions, that port is locked even after the program
ceased to be executed. The port itself cannot be used for, sometimes,
several minutes until the kernel finally unlocks the unused port.
This patch uses a special flag which avoids this issue to happen,
the port in question is flagged to be freed right away after its
host server is terminated. This Eases the long process of starting and
terminating the same server over and over for debugging purposes.