libs/neuronet: Fixed a very nasty bug that happened after a while of use.
The bug was only happening after a certain amount of client connected.
It was due to stray status elements being still in the statuses buffer
_after_ the slave containing them was cleaned. This was fixed by
creating a new function which when called (at disconnection),
purges the statuses buffer from all the elements for a given slave,
ensuring that no further call to a no longer existing slave are attempted.