Implemented GLX Full support.
To activate, at configure time pass --with-lowlib=glx.

This implementation adds full glx support which permit
openGL functions to be used with our library. Since
this is a double buffered process, the function
Neuro_Flip has to be used to flush all changes to
the display buffer. Otherwise, nearly all standard
initialization methods stay the same for starting
a glx context.