misc/arg: On multiple options we now call the callback accordingly.
Before, if we had say the option -v (for verbose) and we added
more than just one : -vvvv. The callback was only called once
thus it was impossible for the end program to count the number
of 'v' used in the option. This commit fixes this issue and calls
the callback for every options it encounter. So for example, if
the option : -vvvv is used, the callback will be called four times.