libs/neuronet server: Attempted to fix a weird client disconnect bug.
This bug happens on a server instance, especially when the client pipe
is broken down (remote client closes the connection unexpectly). Not
all the time though, just from time to time. This fixes a segmentation
fault where the slave instance could not find the underlying master
pointer. This kind of fixes the consequence of the problem but does
not fix the cause. There has to be somewhere that does not correctly
allocate or set either the master or frees only partially a slave
struct.