memory/ebuf: Optimized the allocation of buffers greatly.
Previously, every time the cleaning function SCleanEBuf was called, the
data itself was being freed. This patch changes this behavior by
allocating every pointers in the overhead buffer and reuses any scleaned
objects. This drastically changes the amount of allocations being done
when using this module. However, this means that we no longer support
arbitrary size for the elements pushed to the buffer. Code was added to
safeguard against this.